Fix deprecated

+++ src/main/java/org/traccar/protocol/IotmProtocolDecoder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 - 2021 Anton Tananaev (anton@traccar.org)
+ * Copyright 2020 - 2022 Anton Tananaev (anton@traccar.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -339,7 +339,7 @@ public class IotmProtocolDecoder extends BaseProtocolDecoder {
             buf.readUnsignedByte(); // checksum
 
             MqttMessage response = MqttMessageBuilders.pubAck()
-                    .packetId((short) message.variableHeader().packetId())
+                    .packetId(message.variableHeader().packetId())
                     .build();
 
             if (channel != null) {
